+++ docs/03_addition/natural_less_than_weak_interpretation.html
@@ -8,7 +8,7 @@
 		<b style="color: blue;">prove</b> ZERO_in_R: <b style="color: red;">*</b>B(B in NATURALS <b style="color: red;">&</b> <b style="cursor: pointer;" onclick="window.location.href = '../03_addition/natural_less_than_weak.html';" onmouseover="popup_window(this, 'natural_less_than_weak | 03_addition');" onmouseout="close_popup();">natural_less_than_weak</b>(ZERO, B) <b style="color: red;">-&gt</b> <b style="cursor: pointer;" onclick="window.location.href = '../03_addition/natural_less_than.html';" onmouseover="popup_window(this, 'natural_less_than | 03_addition');" onmouseout="close_popup();">natural_less_than</b>(ZERO, B) <b style="color: red;">|</b> <b style="cursor: pointer;" onclick="window.location.href = '../00_zf/equal_sets.html';" onmouseover="popup_window(this, 'equal_sets | 00_zf');" onmouseout="close_popup();">equal_sets</b>(ZERO, B)){
 			<b style="color: blue;">given</b> <b style="color: red;">|</b>B<b style="color: red;">|</b>;
 			<b style="color: blue;">implies</b> B_in_NATURALS, ZERO_less_than_weak_B;
-			<b style="color: blue;">return</b> <b style="color: green;">swap</b>(<b style="cursor: pointer;" onclick="window.location.href = '../03_addition/natural_zero_minimal.html';" onmouseover="popup_window(this, 'natural_zero_minimal | 03_addition');" onmouseout="close_popup();">natural_zero_minimal</b>(B)(B_in_NATURALS));
+			<b style="color: blue;">return</b> <b style="cursor: pointer;" onclick="window.location.href = '../03_addition/natural_zero_minimal.html';" onmouseover="popup_window(this, 'natural_zero_minimal | 03_addition');" onmouseout="close_popup();">natural_zero_minimal</b>(B)(B_in_NATURALS);
 		}
 		ZERO_in_R = R_def(ZERO)(zero_is_natural, ZERO_in_R);
